Use unobstrusive interface Persistable
This makes the purpose of this entity much clearer. Data that comes from a bulk import often as an ID, so having Spring Data check the ID is probably not what we expect.

So basically the idea is to check on entity level (our domain) wether stuff is new or not and not delegating that to a service or repository.

Spring Data documentation has a whole section about it: https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#jpa.entity-persistence.saving-entites.